Refactor into utilities
