+++ apps/learn-card-browser-extension/src/content/main.ts
@@ -1,4 +1,5 @@
 import type { CredentialCandidate, ExtensionMessage } from '../types/messages';
+import { extractExchangeUrlFromLink, getExtractorProtocols } from '../utils/links';
 
 // Minimal VC shape for type guard usage
 type VerifiableCredential = {
@@ -13,54 +14,7 @@ let observer: MutationObserver | null = null;
 let listenersAttached = false;
 let locHookInstalled = false;
 
-// Configurable extractors for custom-scheme credential request links
-// Each entry can specify protocols (schemes) and parameter keys to try for the actual HTTP(S) URL.
-// Extendable to hosts in the future if needed.
-type LinkExtractor = {
-  protocols?: string[];
-  hosts?: string[];
-  paramKeys: string[];
-};
-
-const LINK_EXTRACTORS: readonly LinkExtractor[] = [
-  {
-    // Examples: msprequest://request?vc_request_url=...
-    //           dccrequest://request?vc_request_url=...
-    protocols: ['msprequest', 'dccrequest', 'https', 'asuprequest'],
-    paramKeys: ['vc_request_url', 'request_url', 'exchange_url', 'vc_url']
-  }
-];
-
-const DEFAULT_PARAM_KEYS = ['vc_request_url', 'request_uri', 'exchange_url', 'vc_url'];
-
-const getExtractorProtocols = () =>
-  Array.from(new Set(LINK_EXTRACTORS.flatMap((e) => e.protocols ?? [])));
-
-const extractExchangeUrlFromLink = (href: string): string | null => {
-  try {
-    const u = new URL(href);
-    const proto = u.protocol.replace(':', '');
-    const host = u.hostname;
-    const extractor = LINK_EXTRACTORS.find(
-      (e) => (e.protocols && e.protocols.includes(proto)) || (e.hosts && e.hosts.includes(host))
-    );
-    const keys = Array.from(new Set([...(extractor?.paramKeys ?? []), ...DEFAULT_PARAM_KEYS]));
-    for (const k of keys) {
-      const val = u.searchParams.get(k);
-      if (val) {
-        try {
-          const dec = decodeURIComponent(val);
-          return dec;
-        } catch {
-          return val;
-        }
-      }
-    }
-    return null;
-  } catch {
-    return null;
-  }
-};
+// link extraction helpers are now centralized in ../utils/links
 
 const debounce = (fn: () => void, wait = 200) => {
   let t: number | undefined;
